+++ tests/test_collections.py
@@ -1,6 +1,7 @@
 import os.path
 import tempfile
 import unittest
+from numpy.random import default_rng
 
 from imagedata import Series, Study, Patient, Cohort
 from imagedata.formats import UnknownInputError
@@ -66,6 +67,32 @@ class TestStudy(unittest.TestCase):
             si.write(os.path.join(d, '2'), opts={'keep_uid': True})
             study = Study(d, input_format='dicom', opts={'split_acquisitions': True})
 
+    def test_anonymize_dicom_study(self):
+        study = Study('data/dicom/cohort.zip?cohort/P2/S1')
+        anon_study = study.anonymize()
+        with tempfile.TemporaryDirectory() as d:
+            anon_study.write(d, keep_uid=True)
+            study1 = Study(d)
+            self.assertEqual(len(study), len(study1))
+            for uid in range(len(study)):
+                self.assertEqual(study[uid].shape, study1[uid].shape)
+
+    def test_anonymize_non_dicom_study(self):
+        rng = default_rng()
+        series1 = Series(rng.standard_normal(24).reshape((2,3,4))*100, dtype=int)
+        series2 = Series(rng.standard_normal(210).reshape((5,6,7))*100, dtype=int)
+        series2.studyInstanceUID = series1.studyInstanceUID
+        study = Study({'1': series1, '2': series2})
+        study.studyInstanceUID = series1.studyInstanceUID
+
+        anon_study = study.anonymize()
+        with tempfile.TemporaryDirectory() as d:
+            anon_study.write(d, keep_uid=True, formats='dicom')
+            study1 = Study(d)
+            self.assertEqual(len(study), len(study1))
+            for uid in range(len(study)):
+                self.assertEqual(study[uid].shape, study1[uid].shape)
+
 
 class TestPatient(unittest.TestCase):
 
@@ -85,6 +112,15 @@ class TestPatient(unittest.TestCase):
         with tempfile.TemporaryDirectory() as d:
             patient.write(d)
 
+    def test_anonymize_patient(self):
+        patient = Patient('data/dicom/cohort.zip?cohort/P2')
+        anon_patient = patient.anonymize()
+        with tempfile.TemporaryDirectory() as d:
+            anon_patient.write(d)
+            patient1 = Patient(d)
+            self.assertEqual(len(patient), len(patient1))
+            self.assertEqual('ANONYMOUS', patient1.patientName)
+
 
 class TestCohort(unittest.TestCase):
 
@@ -102,6 +138,13 @@ class TestCohort(unittest.TestCase):
         cohort = Cohort('data/dicom/cohort.zip')
         with tempfile.TemporaryDirectory() as d:
             cohort.write(d)
+
+    def test_anonymize_cohort(self):
+        cohort = Cohort('data/dicom/cohort.zip?cohort')
+        anon_cohort = cohort.anonymize()
+        with tempfile.TemporaryDirectory() as d:
+            anon_cohort.write(d)
+            cohort1 = Cohort(d)
 
 
 if __name__ == '__main__':